vcap procs must fail when it couldn't find the config from doozer
comment from phil:
"get_component_config can also be used by the cloud_controller when
managing other components.

I think this would be fine though, because the component config should
be there regardless and it's fine for the CC to throw an exception if it
cannot find another components config.
"