Add ORACLE_URL when there is a single oracle service configured
This commit also adds a JDBC_ORACLE_URL variable that contains the
connection string for th JDB Thin driver, which seems to be more
commonly used that the format used by the other relational dbs.

I have choosen Thin over OCI because Thin works even without the
Oracle client libraries on the DEA and seems to have similar performance
characteristics as OCI.

ORACLE_URL is now also a contender for DATABASE_URL, which will only
be set if exactly one of MYSQL_URL, POSTGRESQL_URL and ORACLE_URL
is set.

Ref http://bugs.activestate.com/show_bug.cgi?id=97520